# Conflicts:
#	src/main/java/com/itgirls/bank_system/dto/DepositDto.java
#	src/main/java/com/itgirls/bank_system/model/Deposit.java

+++ src/main/java/com/itgirls/bank_system/service/UserServiceImpl.java
@@ -10,6 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -42,22 +43,45 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public List<UserDto> getAllUsers () {
+    public List<UserDto> getAllUsers() {
         return userRepository.findAll().stream().map(this::convertUserToDto).collect(Collectors.toList());
     }
 
     @Override
-    public UserDto updateUser(UserUpdateDto userUpdateDto) {
-        User newUser = userRepository.findUserById(userUpdateDto.getId());
-        String passwordEncode = passwordEncoder.encode(userUpdateDto.getPassword());
-        newUser.setPassword(passwordEncode);
-        newUser.setRole(userUpdateDto.getRole());
-        newUser.setName(userUpdateDto.getName());
-        newUser.setSurname(userUpdateDto.getSurname());
-        newUser.setEmail(userUpdateDto.getEmail());
-        newUser.setAccounts(userUpdateDto.getAccounts());
-        userRepository.save(newUser);
-        return convertUserToDto(newUser);
+    @Transactional
+    public UserDto updateUser(UserUpdateDto userUpdateDto, boolean usePatch) {
+        User user = userRepository.findById(userUpdateDto.getId()).
+                orElseThrow(() -> new RuntimeException("Пользователь не найден"));
+        if (!usePatch || userUpdateDto.getPassword() != null) {
+            String passwordEncode = passwordEncoder.encode(userUpdateDto.getPassword());
+            user.setPassword(passwordEncode);
+        }
+        if (!usePatch || userUpdateDto.getRole() != null) {
+            user.setRole(userUpdateDto.getRole());
+        }
+        if (!usePatch || userUpdateDto.getName() != null) {
+            user.setName(userUpdateDto.getName());
+        }
+        if (!usePatch || userUpdateDto.getSurname() != null) {
+            user.setSurname(userUpdateDto.getSurname());
+        }
+        if (!usePatch || userUpdateDto.getEmail() != null) {
+            user.setEmail(userUpdateDto.getEmail());
+        }
+        if (!usePatch && userUpdateDto.getAccounts() != null) {
+            user.setAccounts(userUpdateDto.getAccounts());
+        }
+        if (!usePatch && userUpdateDto.getDeposits() != null) {
+            user.setDeposits(userUpdateDto.getDeposits());
+        }
+        if (!usePatch && userUpdateDto.getLoans() != null) {
+            user.setLoans(userUpdateDto.getLoans());
+        }
+        if (usePatch && userUpdateDto.getTransactions() != null) {
+            user.setTransactions(userUpdateDto.getTransactions());
+        }
+        userRepository.save(user);
+        return convertUserToDto(user);
     }
 
     @Override
@@ -67,11 +91,18 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(Long id) {
-        userRepository.deleteById(id);
+    public String deleteUser(Long id) {
+        if (userRepository.existsById(id)) {
+            try {
+                userRepository.deleteById(id);
+                return "Пользователь с идентификаторо м" + id + " успешно удален";
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        } else return "Пользователь с идентификатором " + id + " не найден";
     }
 
-    private UserDto convertUserToDto (User user) {
+    private UserDto convertUserToDto(User user) {
         return UserDto.builder()
                 .id(user.getId())
                 .email(user.getEmail())
@@ -79,6 +110,9 @@ public class UserServiceImpl implements UserService {
                 .name(user.getName())
                 .surname(user.getSurname())
                 .accounts(user.getAccounts())
+                .deposits(user.getDeposits())
+                .loans(user.getLoans())
+                .transactions(user.getTransactions())
                 .build();
     }
 }
